board

+++ carbonlease/src/main/java/com/kh/board/model/dao/BoardMapper.java
@@ -25,7 +25,7 @@ public interface BoardMapper {
 	
 	int replyCount(Long boardNo);
 	
-	Long boardUpdateForm();
+	int boardUpdateForm();
 	
 	int insertClBoard(BoardDTO boardVO);
