+++ carbonlease/src/main/java/com/kh/board/model/service/BoardService.java
@@ -15,7 +15,7 @@ public interface BoardService {
 
 	int boardReplyInsert(ReplyInsertVO riVO);
 
-	Long boardUpdateForm();
+	int boardUpdateForm();
 	
 	int insertBoard(BoardDTO boardVo);
